Hangman guessing now functionals.
You also get clues back with a crude thing. I now need a way to output
to the console and make sure the game plays.

Check winner and loser as well. Will need fail count too.

Also need a tester for to check the percentage wins and the average
moves to win.

Finally load guess from a file.